Build and test with multiple versions of Go
Currently, the Pulsar Go Client is only tested against Go version 1.12. The motivation for this change is to help bring confidence that the Go client is compatible with multiple versions of Go, all being tested with each PR.

I've modified the two GitHub Action workflows that are a part of this repo with a [matrix](https://help.github.com/en/actions/reference/workflow-syntax-for-github-actions#jobsjob_idstrategymatrix) such that we test against multiple versions of Go.

- [X] Make sure that the change passes the CI checks.