Removed blocking queue iterator (#301)
## Motivation
The blocking queue iterator method is not really thread safe in the presence of other read/write operations on the queue. While it makes a snapshot of what's readable at one point, if there are Poll() operations, it might be ending up reading new items before the new ones.

For that, it's better to make a copy of the items in the queue, also considering that iterating over this queue is only done on reconnections and it's not performance sensitive.