Do not allocate MessageIDs on the heap
Passing a function parameter by pointer (or writing a pointer into a
channel) will cause the Go escape analysis to allocate the parameter on
the heap.

This change passes messageID struct instances by value instead of by
pointer; this keeps messageID structs on the stack.

Each message produced or consumed by the library is associated with
a MessageID; keeping instances of the MessageID structure on the stack
reduces heap allocation and associated GC cost.

Signed-off-by: Daniel Ferstay <dferstay@splunk.com>