Update default router to switch partition on all batching thresholds #382 (#383)
Master Issue: #382 

### Motivation


The default router only switches partition when the max delay to publish has elapsed. This PR is about switching as soon as one of the threshold (max number of messages, max number of bytes, max delay) is reached.

### Modifications

The default router has been updated to switch to one of the threshold. To prevent cyclic dependencies between the packages, it has been moved from internal to pulsar package.

### Verifying this change

- [X ] Make sure that the change passes the CI checks.

*(Please pick either of the following options)*

This change added unit tests to verify all cases

### Does this pull request potentially affect one of the following parts:

*If `yes` was chosen, please highlight the changes*

  - Dependencies (does it add or upgrade a dependency): no
  - The public API: no
  - The schema: no
  - The default values of configurations: no
  - The wire protocol: no

### Documentation

  - Does this pull request introduce a new feature? yes
  - If yes, how is the feature documented? GoDocs