Add metric for internal publish latency (#397)
The current publish latency metric measures the time when send is called until the message is acked. The metric can be skewed if batching is enabled. 

For example, if we call send the client will hold that message until x number of messages are collected or an interval has passed. This new metric measures the time when we send the data to the connection until we get an ack from the broker. This cuts out any batch skewing and provides a more accurate latency of what the client sees at the connection level.