Improve error log for frame size too big and maxMessageSize (#459)
### Motivation
This PR is to improve error log for `frame size too big` on the consumer. We have seen a number of frame size too big error and I would like to trace what the exact maxMessageSize is set to that is not default. Debug log was not able to be enabled on the production system.


### Modifications

Log improvement mostly