Add methods to deal with string message ids.
If we want to deal with string message ids, we need methods for being
able to create them from either a string or a using the individual parts
of the message id.

This exposes some constructors for this.

Tests are needed for the string message parsing, but it is taking from
pulsarctl where it is well used