Fix data race while accessing connection in partitionConsumer
The partitionConsumer maintains a few internal go-routines, two of which
access the underlying internal.Connection.  The main runEvenstLoop()
go-routine reads the connection field while a separate go-routine is used
to detect connnection loss, initiate reconnection, and sets the connection.

Previously, access to the conn field was not synchronized.

Now, the conn field is read and written atomically; resolving the data race.

Signed-off-by: Daniel Ferstay <dferstay@splunk.com>