Add jitter to exponential backoff (#556) (#557)
Co-authored-by: Denis Vergnes <dvergnes@users.noreply.github.com>

Fixes #556 


### Motivation
When reconnecting to a broker, an exponential backoff strategy is used. However, there is no randomness involved which means in case of a general outage impacting several clients, they will try to reconnect at the same time.

### Modifications

The modification consists of adding a jitter that adds 20% of delay after the delay is doubled on each retry.