Fix the sink and source tests (streamnative/pulsarctl#411)
* Fix the sink and source tests
---

Fixes streamnative/pulsarctl#406

*Motivation*

Currently, the sink and source tests depend on the kafka and
mysql connectors. And those connectors need download from the
website.
Another thing is we will change the image version when there
has a new release. Current testing steps is hard to change to
the new version, we need to update all the connectors used in
the test, that will introduce unnecessary operations.
In the pulsar-all image, we already have the connectors so we
can use them directly. So I refactor the sink and source tests
to use the connectors in the pulsar image.
Using the data-generator connectors for testing the sink and source
apis in the pulsarctl. The data-generator connectors doesn't need to
depend on other services, it's easier to use in the test.

*Modifications*

- fix the sink and source tests
- update the test scripts to allow test can run by `run-integration-test.sh`
- simplify the test steps
- update the test workflows

*Verify this change*

Please pick either of following options.

- Pass the tests