Fix logic of command for sendError (#622)
### Motivation


![image](https://user-images.githubusercontent.com/20965307/135020293-06cb72cc-5ed9-4bc5-a7ba-3909da57a8a6.png)


As shown in the figure above, the `ServerError` returned by the broker is `UnknownError` when the client receives it. In fact, we handled the wrong command here. Here we should deal with `CommandSendError` instead of `CommandError`. Correspondingly, we should deal with the `listener` map used to cache the producer instead of the corresponding `pendingRequest` map.