[ClientOptions] Allow callers to specify custom metric name qualifiers
This change exposes the full metric name qualification supported in the
underlying prometheus client library by allowing metric names to be
qualified by both a "namespace" and "subsystem".  See the documentation for
the prometheus metric options structure:
https://pkg.go.dev/github.com/prometheus/client_golang@v1.11.0/prometheus#Opts

This is helpful to disambiguate metrics when there is more than one
component in a distributed system that makes use of pulsar-client-go.

Signed-off-by: Daniel Ferstay <dferstay@splunk.com>