[oauth2] Remove oauth2 go.mod and go.sum (#802)
* remove oauth2 module

This removes the module definition in the oauth2 subdirectory.  Since the oauth2 module
is not released independently of the main module, it can/should be treated as a normal
module subdirectory.

Signed-off-by: Paul Gier <paul.gier@datastax.com>

* update tests for newer golang oauth2 dependency

Signed-off-by: Paul Gier <paul.gier@datastax.com>

* oauth2: add note to readme about possible module error

Signed-off-by: Paul Gier <paul.gier@datastax.com>

Signed-off-by: Paul Gier <paul.gier@datastax.com>

### Motivation

When the oauth2 sub-directory was added in PR #313 there wasn't any justification given (AFAICT) for creating a separate sub-module instead of just treating the oauth2 sources as a normal directory.  Since the oauth2 module does not have a separate release cycle, treating it as a sub-module just adds unnecessary complexity.

### Modifications

Removed `go.mod` and `go.sum` from the oauth2 sub-directory.  Updated the main `go.mod` and `go.sum` to included the necessary dependencies.

The main module contained a newer version of `golang.org/x/oauth2` dependency than the oauth2 sub-module, so I had to update some of the tests to match the output of the newer dependency.  This type of dependency mismatch will be avoided in the future using a single module defined in the root directory.