[feat] Support partitioned topic reader (#1178)
Master Issue: #1177

### Motivation

Currently, there is an issue with the reader implementation. If the reader is creating, it won't get the topic metadata from the topic. The reader can only read messages from a single topic. If the topic is a partitioned topic, the reader won't know that and will try to create a non-partition topic with the same name. And it will lead to this issue: https://github.com/apache/pulsar/issues/22032

### Modifications

- Support partitioned topic reader

(cherry picked from commit 3b9b1f8895d8924ec98db4612806b9871f1d135b)