[Issue 1272][connection] Attempt to avoid deadlock during reconnection (#1273)
Fixes #1272 

### Motivation

Producers and consumers register themselves to the connection to be notified when it gets closed. Even though the callback push the events in a channel, it can get stuck and the connection pool is locked which prevents any other caller to get a connection.

### Modifications

This PR makes sure that the callback never blocks.