[Issue #1418] Fix: Properties not getting consistently set on pulsaradmin subscription message responses (#1419)
### Motivation

The properties on messages returned by PeekMessages is not returned consistently as described in #1418 

This change makes sure we set all properties from headers before continuing on, so the returned message should consistently contain all properties.

Currently the properties returned is inconsistent on repeated calls, even if the message returned is the same each time.

### Modifications

I've made it so all headers are processed / al properties are set on every message, rather than exiting the loop early which can cause inconsistent results.